Fixes Forge Microblock ID Conflict
The Chisel mod tries to register several blocks with the same name under
different IDs.  The two that are in conflict are Voidstone and Factory
blocks.  These are purely aesthetic and are only used for decoration in
the Chisel mod, and therefore disabling them to allow Forge Microblocks to
work isn't a problem