Fixes multiple mods' duplicate metal ores
Different mods need the same metals for recipes (E.G. copper), but they all
include their own versions of the ores.  This will remove most of the duplicate
ores so that there's only one version of each metal.  This is okay because most
recipes will still work with other versions of the metals, and if they don't,
you can use the Chisel mod to change the blocks of metal from one mod to
another.